remove unit testing, and change e2e test to test
