+++ .husky/pre-commit
@@ -3,4 +3,3 @@
 
 yarn lint
 yarn test
-yarn test:e2e
