--- e2e/auth/auth.change-password.e2e-spec.ts
@@ -1,179 +1,0 @@
-import { HttpStatus, INestApplication } from '@nestjs/common';
-import { Test } from '@nestjs/testing';
-import request from 'supertest';
-import faker from '@faker-js/faker';
-import { UserDocument } from 'src/user/user.schema';
-import { UserService } from 'src/user/user.service';
-import { RoleService } from 'src/role/role.service';
-import { AuthService } from 'src/auth/auth.service';
-import { RoleDocument } from 'src/role/role.schema';
-import { IUserDocument } from 'src/user/user.interface';
-import { E2E_AUTH_CHANGE_PASSWORD_URL } from './auth.constant.e2e';
-import { ENUM_REQUEST_STATUS_CODE_ERROR } from 'src/request/request.constant';
-import { ENUM_USER_STATUS_CODE_ERROR } from 'src/user/user.constant';
-import { Types } from 'mongoose';
-import { ENUM_AUTH_STATUS_CODE_ERROR } from 'src/auth/auth.constant';
-import { CoreModule } from 'src/core/core.module';
-import { RouterCommonModule } from 'src/router/router.common.module';
-import { RouterModule } from '@nestjs/core';
-
-describe('E2E Change Password', () => {
-    let app: INestApplication;
-    let userService: UserService;
-    let authService: AuthService;
-    let roleService: RoleService;
-
-    const password = `@!${faker.random
-        .alphaNumeric(5)
-        .toLowerCase()}${faker.random.alphaNumeric(5).toUpperCase()}`;
-    const newPassword = `@!${faker.random
-        .alphaNumeric(5)
-        .toLowerCase()}${faker.random.alphaNumeric(5).toUpperCase()}`;
-
-    let user: UserDocument;
-
-    let accessToken: string;
-    let accessTokenNotFound: string;
-
-    beforeAll(async () => {
-        const modRef = await Test.createTestingModule({
-            imports: [
-                CoreModule,
-                RouterCommonModule,
-                RouterModule.register([
-                    {
-                        path: '/',
-                        module: RouterCommonModule,
-                    },
-                ]),
-            ],
-        }).compile();
-
-        app = modRef.createNestApplication();
-        userService = app.get(UserService);
-        authService = app.get(AuthService);
-        roleService = app.get(RoleService);
-
-        const role: RoleDocument = await roleService.findOne({
-            name: 'user',
-        });
-
-        const passwordHash = await authService.createPassword(password);
-
-        user = await userService.create({
-            firstName: faker.name.firstName(),
-            lastName: faker.name.lastName(),
-            password: passwordHash.passwordHash,
-            passwordExpired: passwordHash.passwordExpired,
-            salt: passwordHash.salt,
-            email: faker.internet.email(),
-            mobileNumber: faker.phone.phoneNumber('62812#########'),
-            role: `${role._id}`,
-        });
-
-        const userPopulate = await userService.findOneById<IUserDocument>(
-            user._id,
-            {
-                populate: {
-                    role: true,
-                    permission: true,
-                },
-            }
-        );
-
-        const map = await authService.mapLogin(userPopulate);
-        const payload = await authService.createPayload(map, false);
-        const payloadNotFound = {
-            ...payload,
-            _id: `${new Types.ObjectId()}`,
-        };
-
-        accessToken = await authService.createAccessToken(payload);
-        accessTokenNotFound = await authService.createAccessToken(
-            payloadNotFound
-        );
-        await app.init();
-    });
-
-    it(`PATCH ${E2E_AUTH_CHANGE_PASSWORD_URL} Error Request`, async () => {
-        const response = await request(app.getHttpServer())
-            .patch(E2E_AUTH_CHANGE_PASSWORD_URL)
-            .send({
-                oldPassword: '123123',
-                newPassword: '123',
-            })
-            .set('Authorization', `Bearer ${accessToken}`);
-
-        expect(response.status).toEqual(HttpStatus.UNPROCESSABLE_ENTITY);
-        expect(response.body.statusCode).toEqual(
-            ENUM_REQUEST_STATUS_CODE_ERROR.REQUEST_VALIDATION_ERROR
-        );
-    });
-
-    it(`PATCH ${E2E_AUTH_CHANGE_PASSWORD_URL} Not Found`, async () => {
-        const response = await request(app.getHttpServer())
-            .patch(E2E_AUTH_CHANGE_PASSWORD_URL)
-            .send({
-                oldPassword: password,
-                newPassword,
-            })
-            .set('Authorization', `Bearer ${accessTokenNotFound}`);
-
-        expect(response.status).toEqual(HttpStatus.NOT_FOUND);
-        expect(response.body.statusCode).toEqual(
-            ENUM_USER_STATUS_CODE_ERROR.USER_NOT_FOUND_ERROR
-        );
-    });
-
-    it(`PATCH ${E2E_AUTH_CHANGE_PASSWORD_URL} Old Password Not Match`, async () => {
-        const response = await request(app.getHttpServer())
-            .patch(E2E_AUTH_CHANGE_PASSWORD_URL)
-            .send({
-                oldPassword: 'as1231dAA@@!',
-                newPassword,
-            })
-            .set('Authorization', `Bearer ${accessToken}`);
-
-        expect(response.status).toEqual(HttpStatus.BAD_REQUEST);
-        expect(response.body.statusCode).toEqual(
-            ENUM_AUTH_STATUS_CODE_ERROR.AUTH_PASSWORD_NOT_MATCH_ERROR
-        );
-    });
-
-    it(`PATCH ${E2E_AUTH_CHANGE_PASSWORD_URL} New Password must different with old password`, async () => {
-        const response = await request(app.getHttpServer())
-            .patch(E2E_AUTH_CHANGE_PASSWORD_URL)
-            .send({
-                oldPassword: password,
-                newPassword: password,
-            })
-            .set('Authorization', `Bearer ${accessToken}`);
-
-        expect(response.status).toEqual(HttpStatus.BAD_REQUEST);
-        expect(response.body.statusCode).toEqual(
-            ENUM_AUTH_STATUS_CODE_ERROR.AUTH_PASSWORD_NEW_MUST_DIFFERENCE_ERROR
-        );
-    });
-
-    it(`PATCH ${E2E_AUTH_CHANGE_PASSWORD_URL} Success`, async () => {
-        const response = await request(app.getHttpServer())
-            .patch(E2E_AUTH_CHANGE_PASSWORD_URL)
-            .send({
-                oldPassword: password,
-                newPassword,
-            })
-            .set('Authorization', `Bearer ${accessToken}`);
-
-        expect(response.status).toEqual(HttpStatus.OK);
-        expect(response.body.statusCode).toEqual(HttpStatus.OK);
-    });
-
-    afterAll(async () => {
-        try {
-            await userService.deleteOneById(user._id);
-        } catch (e) {
-            console.error(e);
-        }
-        await app.close();
-    });
-});
