--- e2e/jest.json
@@ -1,30 +1,0 @@
-{
-    "testTimeout": 1000,
-    "verbose": true,
-    "rootDir": "../",
-    "modulePaths": [
-        "./"
-    ],
-    "testEnvironment": "node",
-    "testRegex": "e2e/.*\\.(e2e-spec).(ts|tsx)$",
-    "collectCoverage": true,
-    "collectCoverageFrom" : ["e2e/**/*.{tsx,ts}", "!**/test/**/*.{js,jsx,tsx,ts}", "!**/src/**/*.{js,jsx,tsx,ts}", "!**/dist/**", "!**/node_modules/**", "!**/vendor/**"],
-    "coveragePathIgnorePatterns": [
-        "node_modules", 
-        "vendor",
-        "src"
-    ],
-    "coverageReporters": ["clover", "json", "lcov", "text"],
-    "coverageDirectory": "e2e-coverage",
-    "moduleFileExtensions": [
-        "js", 
-        "jsx", 
-        "ts", 
-        "tsx", 
-        "json", 
-        "node"
-    ],
-    "transform": {
-        "^.+\\.tsx?$": "ts-jest"
-    }
-}
