+++ package.json
@@ -15,26 +15,23 @@
   "scripts": {
     "prebuild": "rimraf dist",
     "build": "nest build",
-    "format": "prettier --write '{src,e2e,test}/**/*.ts'",
+    "format": "prettier --write '{src,test}/**/*.ts'",
     "format:src": "prettier --write src/**/*.ts",
     "format:test": "prettier --write test/**/*.ts",
-    "format:e2e": "prettier --write e2e/**/*.ts",
     "rollback": "nestjs-command remove:user && nestjs-command remove:role && nestjs-command remove:permission",
     "migrate": "nestjs-command insert:permission && nestjs-command insert:role && nestjs-command insert:user",
     "start": "nest start",
     "start:dev": "nest start --watch",
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/main",
-    "lint": "eslint --ext .ts,.tsx '{src,e2e,test}/**/*.ts' --no-error-on-unmatched-pattern",
-    "lint:fix": "eslint --ext .ts,.tsx '{src,e2e,test}/**/*.ts' --fix --no-error-on-unmatched-pattern",
+    "lint": "eslint --ext .ts,.tsx '{src,test}/**/*.ts' --no-error-on-unmatched-pattern",
+    "lint:fix": "eslint --ext .ts,.tsx '{src,test}/**/*.ts' --fix --no-error-on-unmatched-pattern",
     "lint:test": "eslint --ext .ts,.tsx test/**/*.ts --no-error-on-unmatched-pattern",
-    "lint:e2e": "eslint --ext .ts,.tsx e2e/**/*.ts --no-error-on-unmatched-pattern",
     "lint:src": "eslint --ext .ts,.tsx src/**/*.ts --no-error-on-unmatched-pattern",
     "test": "jest --config test/jest.json --detectOpenHandles --passWithNoTests --forceExit",
     "test:watch": "jest --config test/jest.json --watch",
     "test:cov": "jest --config test/jest.json --coverage",
     "test:debug": "node --config test/jest.json --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "jest --config e2e/jest.json --detectOpenHandles --passWithNoTests --forceExit",
     "prepare": "husky install"
   },
   "dependencies": {
