+++ test/auth/auth.public.spec.ts
@@ -3,12 +3,13 @@ import { Test } from '@nestjs/testing';
 import request from 'supertest';
 import faker from '@faker-js/faker';
 import { UserService } from 'src/user/user.service';
-import { E2E_AUTH_PUBLIC_SIGN_UP_URL } from './auth.constant.e2e';
+import { E2E_AUTH_PUBLIC_SIGN_UP_URL } from './auth.constant';
 import { ENUM_REQUEST_STATUS_CODE_ERROR } from 'src/request/request.constant';
 import { ENUM_USER_STATUS_CODE_ERROR } from 'src/user/user.constant';
 import { CoreModule } from 'src/core/core.module';
 import { RouterPublicModule } from 'src/router/router.public.module';
 import { RouterModule } from '@nestjs/core';
+import { connection } from 'mongoose';
 
 describe('E2E Public', () => {
     let app: INestApplication;
@@ -62,6 +63,8 @@ describe('E2E Public', () => {
         expect(response.body.statusCode).toEqual(
             ENUM_REQUEST_STATUS_CODE_ERROR.REQUEST_VALIDATION_ERROR
         );
+
+        return;
     });
 
     it(`POST ${E2E_AUTH_PUBLIC_SIGN_UP_URL} Sign Up Success`, async () => {
@@ -84,6 +87,8 @@ describe('E2E Public', () => {
         expect(response.body.statusCode).toEqual(
             ENUM_USER_STATUS_CODE_ERROR.USER_EXISTS_ERROR
         );
+
+        return;
     });
 
     it(`POST ${E2E_AUTH_PUBLIC_SIGN_UP_URL} Sign Up Email Exist`, async () => {
@@ -99,6 +104,8 @@ describe('E2E Public', () => {
         expect(response.body.statusCode).toEqual(
             ENUM_USER_STATUS_CODE_ERROR.USER_EMAIL_EXIST_ERROR
         );
+
+        return;
     });
 
     it(`POST ${E2E_AUTH_PUBLIC_SIGN_UP_URL} Sign Up Mobile Number Exist`, async () => {
@@ -114,6 +121,8 @@ describe('E2E Public', () => {
         expect(response.body.statusCode).toEqual(
             ENUM_USER_STATUS_CODE_ERROR.USER_MOBILE_NUMBER_EXIST_ERROR
         );
+
+        return;
     });
 
     afterAll(async () => {
@@ -125,6 +134,8 @@ describe('E2E Public', () => {
         } catch (e) {
             console.error(e);
         }
+
+        connection.close();
         await app.close();
     });
 });
